+++ itk_dev_shared_components/eflyt/util.py
@@ -1,0 +1,26 @@
+'''General helper funtions'''
+from datetime import date
+
+from selenium import webdriver
+from selenium.webdriver.common.by import By
+from selenium.webdriver.remote.webelement import WebElement
+
+
+def format_date(_date: date) -> str:
+    '''Format date as %d%m%Y'''
+    return _date.strftime("%d%m%Y")
+
+
+def extract_rows(browser: webdriver.Chrome, row_id: str) -> list[WebElement]:
+    '''Get rows by ID
+
+    Args:
+        browser: The browser app to use
+        row_id: The row to extract
+
+    Return:
+        A list of WebElements'''
+    table = browser.find_element(By.ID, row_id)
+    rows = table.find_elements(By.TAG_NAME, "tr")
+    # Remove header row
+    rows.pop(0)