+++ tests/test_eflyt/test_case.py
@@ -1,0 +1,10 @@
+'''Tests related to the Case module'''
+
+import unittest
+
+
+class CaseTest(unittest.testcase):
+    '''Test the Case functionality of Eflyt integration'''
+
+    def test_find_data(self):
+        '''Look for data and verify format'''