+++ tests/test_eflyt/test_login.py
@@ -1,0 +1,10 @@
+'''Tests related to login in Eflyt'''
+
+import unittest
+
+
+class LoginTest(unittest.TestCase):
+    '''Test login functionality in the Eflyt module'''
+
+    def test_login(self):
+        '''Test to see if we can login'''