+++ tests/test_eflyt/test_move_search.py
@@ -1,0 +1,10 @@
+'''Tests related to the Move Search Eflyt module'''
+
+import unittest
+
+
+class MoveSearchTest(unittest.testcase):
+    '''Test the Move Search functionality of Eflyt integration'''
+
+    def test_lookup(self):
+        '''Lookup a CPR and check that it is found'''