+++ tests/test_eflyt/test_super_search.py
@@ -1,0 +1,10 @@
+'''Tests related to the Super Search Eflyt module'''
+
+import unittest
+
+
+class SuperSearchTest(unittest.testcase):
+    '''Test the Super Search functionality of Eflyt integration'''
+
+    def test_lookup(self):
+        '''Lookup a CPR and check that it is found'''